feat: 데스크탑 채팅 ui 고도화
