+++ src/components/ui/NavbarContent.tsx
@@ -1,16 +1,15 @@
 import { useLocation } from "react-router-dom";
 import { ContentCategory } from "./ContentCategory";
+import ChatNavbar from "components/chat/ChatNavbar";
 
 export function NavbarContent() {
   const { pathname } = useLocation();
   const page = pathname.split("/")[1];
   console.log(page);
   switch (page) {
-    case "":
-      return <ContentCategory />;
     case "chat":
-      return <p>chat navbar</p>;
+      return <ChatNavbar />;
     default:
-      return <p>default</p>;
+      return <ContentCategory />;
   }
 }
